add LargeNumberLogic
теперь есть прогрессия цифр, не проводил глубоких тестов но за диапазон не выхожу все работает, думаю для такого проекта этой логики будет достаточно теперь стоит заняться другой частью игры, проект кажется подходит к завершению

+++ Assets/Scripts/Data/RevealShopData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public struct RevealShopData
 {
     [SerializeField] string name;
-    [SerializeField] int revealValue;
+    [SerializeField] float revealValue;
 
-    public int RevealValue { get => revealValue; }
+    public float RevealValue { get => revealValue; }
 }
